Moved CommonResponse to the Common module within provider models to centralize response handling. This simplifies the process of adding new AI providers by providing a consistent response structure. Breaking change: Existing providers may are also have updated response handling to conform to the new changes
